requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
cohere==4.37
google-generativeai==0.3.1
python-dotenv==1.0.0